WIP : nsls2/io/output.py have to work on.gsas and .xye format

+++ nsls2/io/output.py
@@ -34,8 +34,9 @@
 ########################################################################
 
 """
-    This module is for saving the powder x-ray diffraction
-    intensities into  different file formats
+    This module is for saving integrated powder x-ray diffraction
+    intensities into  different file formats.
+    (Output into different file formats, .chi, .dat, .xye, .gsas)
 
 """
 
@@ -45,38 +46,43 @@ import scipy.io
 import os
 
 
-def save_chi(tth, intensity, err, filename, dir_name=None):
+def save_chi(tth, intensity, err, filename, dir_path=None):
     """
-    Save diffraction intensities into .chi file format
+    Save output diffraction intensities into .chi file format
 
     Parameters
     ----------
     tth : ndarray
-        2(theta) values or (Q values)
+        2(theta) values or (Q values) 1XN array
 
     intensity : ndarray
-               intensity values
+        intensity values 1XN array
 
     err : ndarray
-          uncertainty
-
-    path : str
-           directory to save the chi files
+        uncertainty 1XN array
 
     filename : str
-               filename(could be full path) of the .tif file
+        filename of the .tif file
 
-    dir_name : str, optional
+    dir_path : str, optional
         new directory path to save the output data files
-
 
     Returns
     -------
     output : str
              saved file of diffraction intensities in .chi file format
     """
+
     file_base = os.path.splitext(os.path.split(filename)[1])[0]
-    file_path = file_base + '.chi'
+
+    if len(dir_path) > 0 and not os.path.exists(dir_path):
+        raise ValueError('The given path does not exist.')
+    if not os.path.isabs(dir_path):
+        print 'Creating image: ' + os.path.join(os.getcwd(), dir_path, filename)
+        file_path = dir_path + file_base + '.chi'
+    else:
+        #print 'Creating image: ' + os.path.join(dir_path, filename)
+        file_path = file_base + '.chi'
 
     f = open(file_path, 'wb')
     np.savetxt(f, (tth, intensity, err), newline='\n')
@@ -84,9 +90,9 @@ def save_chi(tth, intensity, err, filename, dir_name=None):
     return
 
 
-def save_dat(tth, intensity, err, filename, dir_name=None):
+def save_dat(tth, intensity, err, filename, dir_path=None):
     '''
-    Save diffraction intensities into .dat file format
+    Save output diffraction intensities into .dat file format
 
     Parameters
     ----------
@@ -99,15 +105,11 @@ def save_dat(tth, intensity, err, filename, dir_name=None):
     err : ndarray
         uncertainty
 
-    path : str
-        directory to save the chi files
-
     filename : str
         filename(could be full path) of the .tif file
 
-    dir_name : str, optional
+    dir_path : str, optional
         name of the new directory to save the output data files
-
 
     Returns
     -------
@@ -139,15 +141,11 @@ def save_xye(tth, intensity, err, filename, dir_name=None):
     err : ndarray
         uncertainty
 
-    path : str
-        directory to save the chi files
-
     filename : str
         filename(could be full path) of the .tif file
 
-    dir_name : str, optional
-        name of the new directory to save the output data files
-
+     dir_path : str, optional
+        new directory path to save the output data files
 
     Returns
     -------
@@ -165,7 +163,7 @@ def save_xye(tth, intensity, err, filename, dir_name=None):
     pass
 
 
-def save_gsas(tth, intensity, err, filename, dir_name=None):
+def save_gsas(tth, intensity, filename, err=None dir_name=None):
     """
     Save diffraction intensities into .gsas file format
 
@@ -177,17 +175,17 @@ def save_gsas(tth, intensity, err, filename, dir_name=None):
     intensity : ndarray
         intensity values
 
-    err : ndarray
-        uncertainty
-
     path : str
         directory to save the chi files
 
     filename : str
         filename(could be full path) of the .tif file
 
-    dir_name : str, optional
-        name of the new directory to save the output data files
+    err : ndarray, None
+        uncertainty
+
+     dir_path : str, optional
+        new directory path to save the output data files
 
 
     Returns
@@ -196,3 +194,77 @@ def save_gsas(tth, intensity, err, filename, dir_name=None):
         Saved file of diffraction intensities in .gsas file format
 
     """
+
+ def saveGSAS(self, xrd, filename):
+        '''
+        save diffraction intensity in gsas format
+
+        :param xrd: 2d array with shape (2,len of intensity) or (3, len of intensity), [tthorq, intensity, (unceratinty)]
+        :param filename: str, base file name
+        '''
+        filepath = self.getFilePathWithoutExt(filename) + '.gsas'
+        f = open(filepath, 'wb')
+        f.write(self.config.getHeader(mode='short'))
+        f.write('#### start data\n')
+        if xrd.shape[0] == 3:
+            s = writeGSASStr(os.path.splitext(path)[0], self.gsasoutput, xrd[0], xrd[1], xrd[2])
+        elif xrd.shape[0] == 2:
+            s = writeGSASStr(os.path.splitext(path)[0], self.gsasoutput, xrd[0], xrd[1])
+        f.write(s)
+        f.close()
+        return filepath
+
+def writeGSASStr(name, mode, tth, iobs, esd=None):
+    """
+    Return string of integrated intensities in GSAS format.
+    :param mode: string, gsas file type, could be 'std', 'esd', 'fxye' (gsas format)
+    :param tth: ndarray, two theta angle
+    :param iobs: ndarray, Xrd intensity
+    :param esd: ndarray, optional error value of intensity
+
+    :return:  string, a string to be saved to file
+    """
+    maxintensity = 999999
+    logscale = numpy.floor(numpy.log10(maxintensity / numpy.max(iobs)))
+    logscale = min(logscale, 0)
+    scale = 10 ** int(logscale)
+    lines = []
+    ltitle = 'Angular Profile'
+    ltitle += ': %s' % name
+    ltitle += ' scale=%g' % scale
+    if len(ltitle) > 80:    ltitle = ltitle[:80]
+    lines.append("%-80s" % ltitle)
+    ibank = 1
+    nchan = len(iobs)
+    # two-theta0 and dtwo-theta in centidegrees
+    tth0_cdg = tth[0] * 100
+    dtth_cdg = (tth[-1] - tth[0]) / (len(tth) - 1) * 100
+    if esd == None: mode = 'std'
+    if mode == 'std':
+        nrec = int(numpy.ceil(nchan / 10.0))
+        lbank = "BANK %5i %8i %8i CONST %9.5f %9.5f %9.5f %9.5f STD" % \
+                (ibank, nchan, nrec, tth0_cdg, dtth_cdg, 0, 0)
+        lines.append("%-80s" % lbank)
+        lrecs = [ "%2i%6.0f" % (1, ii * scale) for ii in iobs ]
+        for i in range(0, len(lrecs), 10):
+            lines.append("".join(lrecs[i:i + 10]))
+    if mode == 'esd':
+        nrec = int(numpy.ceil(nchan / 5.0))
+        lbank = "BANK %5i %8i %8i CONST %9.5f %9.5f %9.5f %9.5f ESD" % \
+                (ibank, nchan, nrec, tth0_cdg, dtth_cdg, 0, 0)
+        lines.append("%-80s" % lbank)
+        lrecs = [ "%8.0f%8.0f" % (ii, ee * scale) for ii, ee in zip(iobs, esd) ]
+        for i in range(0, len(lrecs), 5):
+            lines.append("".join(lrecs[i:i + 5]))
+    if mode == 'fxye':
+        nrec = nchan
+        lbank = "BANK %5i %8i %8i CONST %9.5f %9.5f %9.5f %9.5f FXYE" % \
+                (ibank, nchan, nrec, tth0_cdg, dtth_cdg, 0, 0)
+        lines.append("%-80s" % lbank)
+        lrecs = [ "%22.10f%22.10f%24.10f" % (xx * scale, yy * scale, ee * scale) for xx, yy, ee in zip(tth, iobs, esd) ]
+        for i in range(len(lrecs)):
+            lines.append("%-80s" % lrecs[i])
+    lines[-1] = "%-80s" % lines[-1]
+    rv = "\r\n".join(lines) + "\r\n"
+    return rv
+
